+++ IntegrationTests/Support/TestEnvironment.cs
@@ -1,11 +1,18 @@
-﻿namespace IntegrationTests.Support
+﻿using Iov42sdk.Support;
+
+namespace IntegrationTests.Support
 {
     public class TestEnvironment
     {
         static TestEnvironment()
         {
             Environment = System.Environment.GetEnvironmentVariable("IOV42_ENVIRONMENT");
+            DefaultClientSettings = new ClientSettings(Environment)
+                .WithReadDelayForConsistency()
+                .WithRedirectDelay(500);
         }
+
+        public static ClientSettings DefaultClientSettings { get; }
 
         public static string Environment { get; }
     }