+++ Iov42sdk/Models/Health/BrokerHealthStatus.cs
@@ -1,6 +1,6 @@
 ﻿namespace Iov42sdk.Models.Health
 {
-    public class WriteServiceStatus
+    public class BrokerHealthStatus
     {
         public bool CanWrite { get; set; }
     }