+++ Iov42sdk/Models/Health/StoreHealthStatus.cs
@@ -1,6 +1,6 @@
 ﻿namespace Iov42sdk.Models.Health
 {
-    public class ReadServiceStatus
+    public class StoreHealthStatus
     {
         public bool CanRead { get; set; }
         public bool CanWrite { get; set; }