+++ Iov42sdk/Support/EventualConsistency.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Threading.Tasks;
 
 namespace Iov42sdk.Support
@@ -9,10 +10,12 @@ namespace Iov42sdk.Support
 
         private DateTime _lastWriteLimit = DateTime.MinValue;
         private DateTime _start;
+        private bool _writePerformed;
 
         public void StartWriteOperation()
         {
             _start = DateTime.UtcNow;
+            _writePerformed = true;
         }
 
         public void EndWriteOperation()
@@ -20,14 +23,17 @@ namespace Iov42sdk.Support
             var now = DateTime.UtcNow;
             // Save the best estimate when persistence will be complete - crudely scaled
             // based on how long the last write operation took
-            _lastWriteLimit = now.AddMilliseconds(now.Subtract(_start).TotalMilliseconds * 2);
+            var totalMilliseconds = now.Subtract(_start).TotalMilliseconds;
+            Debug.WriteLine(totalMilliseconds);
+            _lastWriteLimit = now.AddMilliseconds(totalMilliseconds);
         }
 
         public async Task ReadOperation()
         {
             // Allow eventual consistency of persistence to kick in
-            if (_lastWriteLimit <= DateTime.UtcNow) 
+            if (_lastWriteLimit <= DateTime.UtcNow || !_writePerformed) 
                 return;
+            _writePerformed = false;
             await Task.Delay(ConsistencyDelay);
             _lastWriteLimit = DateTime.MinValue;
         }